third_party/blink/web_tests/external/wpt/service-workers/service-worker/worker-interception.https-expected.txt

This is a testharness.js-based test.
FAIL Verify a dedicated worker script request issued from a uncontrolled document is intercepted by worker's own service worker. promise_test: Unhandled rejection with value: undefined
PASS Verify a shared worker script request issued from a uncontrolled document is intercepted by worker's own service worker.
FAIL Verify a same-origin worker script served by a service worker succeeds in starting a dedicated worker. promise_test: Unhandled rejection with value: undefined
PASS Verify a same-origin worker script served by a service worker succeeds in starting a shared worker.
PASS Verify a cors worker script served by a service worker fails dedicated worker start.
PASS Verify a cors worker script served by a service worker fails shared worker start.
PASS Verify a no-cors cross-origin worker script served by a service worker fails dedicated worker start.
PASS Verify a no-cors cross-origin worker script served by a service worker fails shared worker start.
PASS Verify subresource requests on a dedicated worker controlled by a service worker.
FAIL Verify subresource requests on a shared worker controlled by a service worker. assert_equals: expected "finish" but got "failure:Error: assert_equals: expected \"This load was successfully intercepted.\" but got \"Unexpected error! Worker is not defined\""
Harness: the test ran to completion.
